requirements.txt
Flask==1.1.1
line-bot-sdk==1.16.0

autopep8==1.5
backoff==1.10.0
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chromedriver-binary==80.0.3987.16.0
Click==7.0
Django==2.2.6
django-bootstrap4==1.1.1
Flask==1.1.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
httplib2==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
modules==1.0.0
numpy==1.18.1
oauth2client==4.1.3
pandas==1.0.1
pep8==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
PyDrive==1.3.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
